+++ example/src/main/kotlin/ua/wwind/example/table/EventTable.kt
@@ -1,12 +1,14 @@
-@file:OptIn(kotlin.time.ExperimentalTime::class)
-package ua.wwind.example
+@file:OptIn(ExperimentalTime::class)
+
+package ua.wwind.example.table
 
 import org.jetbrains.exposed.v1.core.Column
 import org.jetbrains.exposed.v1.core.Table
 import org.jetbrains.exposed.v1.datetime.date
 import org.jetbrains.exposed.v1.datetime.timestamp
+import kotlin.time.ExperimentalTime
 
-object Events : Table("events") {
+object EventTable : Table("events") {
     val id: Column<Int> = integer("id").autoIncrement()
     val title: Column<String> = varchar("title", length = 100)
     val day = date("day")